Feat(PurchaseAmount): Add PurchaseAmount domain and move validation logic inside
