+++ src/main/java/lotto/Validator.java
@@ -6,18 +6,6 @@ import java.util.Set;
 
 public class Validator {
 
-    public void validatePurchaseAmountDivisibleByThousand(int purchaseAmount) {
-        if (purchaseAmount % 1000 != 0) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    public void validatePurchaseAmountPositive(int purchaseAmount) {
-        if (purchaseAmount <= 0) {
-            throw new IllegalArgumentException();
-        }
-    }
-
     public void validateWinningNumbersCount(List<Integer> winningNumbers) {
         if (winningNumbers.size() != 6) {
             throw new IllegalArgumentException();
